% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.mat.R
\docType{data}
\name{prox.mat}
\alias{prox.mat}
\title{Proximity Matrix for Small Area Estimation under Spatial SAR Model}
\format{
The values are numbers in the interval [0,1] containing the proximity of the row and column domains. The sum of the values of each row is equal to 1.
}
\usage{
data(prox.mat)
}
\description{
A data frame containing the proximity values for the 64 regions to simulate Small Area Estimation under Spatial SAR Model using Hierarchical Bayesian Method
}
\keyword{datasets}
